Add files via upload
The proposed system is a distributed home automation system, consists of a server, sensors. Server controls and monitors the various sensors, and can be easily configured to handle more hardware interface module (sensors). This project is based on the construction of a model simulating home automation with different operation modes which can be controlled also by a mobile application. To achieve this objective, a scale house that captures different signals, both digital and analog, has been developed. To approach the house to a real home automation application the variables under study and control are lighting home appliances and other electrical objects.